Move client manager into sylius bridge
