+++ src/Monofony/Bundle/CoreBundle/Recipe/config/packages/monofony_core.yaml
@@ -12,7 +12,7 @@ fos_oauth_server:
     auth_code_class:     App\Entity\OAuth\AuthCode
     service:
         user_provider: sylius.app_user_provider.email_or_name_based
-        client_manager: Monofony\Bundle\CoreBundle\Entity\OAuth\ClientManager
+        client_manager: Monofony\Bridge\Sylius\User\Model\ClientManager
 
 framework:
     assets:
