+++ src/Monofony/Bundle/CoreBundle/Resources/config/services.yaml
@@ -10,5 +10,5 @@ services:
         tags:
             - { name: data_collector, template: '@MonofonyCore/DataCollector/monofony.html.twig', id: monofony_core, priority: -513 }
 
-    Monofony\Bundle\CoreBundle\Entity\OAuth\ClientManager:
+    Monofony\Bridge\Sylius\User\Model\ClientManager:
         $class: '%fos_oauth_server.model.client.class%'
